chore: move benches prelude
