+++ benches/bench.rs
@@ -3,7 +3,7 @@
 use criterion::{criterion_group, criterion_main};
 
 mod benches;
-mod prelude;
+pub(crate) use benches::prelude;
 
 criterion_group!(benches, benches::group);
 criterion_main!(benches);
